Move Requirement class to own file
